# lab-apps-moveis/trabalho-2025-1-cloud-continuousentregation/api/gateway/swagger.yaml
openapi: 3.0.3
info:
  title: API Gateway - Sistema de Entregas
  description: |
    API Gateway para sistema de entregas que integra serviços de:
    - Autenticação de usuários
    - Gerenciamento de pedidos
    - Rastreamento de entregas
    - Notificações e comunicações automáticas (Serverless)
  version: 1.0.0
  contact:
    name: Suporte API
    email: suporte@empresa.com

servers:
  - url: http://localhost:50092
    description: Servidor de desenvolvimento
  - url: http://52.91.225.62:50092
    description: Servidor de produção

tags:
  - name: Auth
    description: Autenticação e autorização de usuários
  - name: Pedidos
    description: Gerenciamento de pedidos
  - name: Rastreamento
    description: Rastreamento de entregas
  - name: Notificacoes
    description: Notificações push e campanhas
  - name: Email
    description: Envio de e-mails automáticos
  - name: Promocoes
    description: Promoções e campanhas
  - name: Segmentacao
    description: Segmentação de usuários

paths:
  /rastreamento/criar:
    post:
      tags: [Rastreamento]
      summary: Criar rastreamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RastreamentoRequest'
      responses:
        '200':
          description: Localização criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  /rastreamento/atualizar:
    put:
      tags: [Rastreamento]
      summary: Atualizar rastreamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RastreamentoRequest'
      responses:
        '200':
          description: Localização atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  /rastreamento/{pedidoId}:
    get:
      tags: [Rastreamento]
      summary: Buscar localização do pedido
      parameters:
        - name: pedidoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Localização encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rastreamento'

  /pedidos:
    post:
      tags: [Pedidos]
      summary: Criar pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoRequest'
      responses:
        '201':
          description: Pedido criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
    get:
      tags: [Pedidos]
      summary: Listar pedidos
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pedido'

  /pedidos/{id}:
    get:
      tags: [Pedidos]
      summary: Buscar pedido por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
    put:
      tags: [Pedidos]
      summary: Atualizar pedido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoUpdateRequest'
      responses:
        '200':
          description: Pedido atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
    delete:
      tags: [Pedidos]
      summary: Cancelar pedido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedido cancelado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  /pedidos/calcular-rota:
    post:
      tags: [Pedidos]
      summary: Calcular rota
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RotaRequest'
      responses:
        '200':
          description: Rota calculada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotaResponse'

  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login realizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuário registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'

  /promocoes:
    put:
      tags: [Promocoes]
      summary: Criar promoção
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Promoção criada
          content:
            application/json:
              schema:
                type: object
    get:
      tags: [Promocoes]
      summary: Listar promoções
      responses:
        '200':
          description: Lista de promoções
          content:
            application/json:
              schema:
                type: object

  /promocoes/enviar:
    post:
      tags: [Promocoes]
      summary: Enviar notificação de promoção
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                mensagem:
                  type: string
      responses:
        '200':
          description: Notificação enviada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  /segmentacao/cidade:
    post:
      tags: [Segmentacao]
      summary: Enviar notificação segmentada por cidade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cidade:
                  type: string
                  example: "São Paulo"
      responses:
        '200':
          description: Notificação enviada para usuários da cidade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  /notificacoes/enviar:
    post:
      tags: [Notificacoes]
      summary: Enviar notificação push
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                mensagem:
                  type: object
                  properties:
                    title:
                      type: string
                    body:
                      type: string
                    data:
                      type: object
      responses:
        '200':
          description: Notificação enviada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  /notificacoes/campanha:
    post:
      tags: [Notificacoes]
      summary: Enviar campanha promocional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    type: string
                mensagem:
                  type: object
                  properties:
                    title:
                      type: string
                    body:
                      type: string
      responses:
        '200':
          description: Campanha enviada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  /email/enviar:
    post:
      tags: [Email]
      summary: Enviar e-mail automático
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to:
                  type: string
                  format: email
                subject:
                  type: string
                text:
                  type: string
      responses:
        '200':
          description: E-mail enviado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

components:
  schemas:
    RastreamentoRequest:
      type: object
      properties:
        id_pedido:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        timestamp:
          type: string
      required: [id_pedido, latitude, longitude]
    Rastreamento:
      type: object
      properties:
        id:
          type: integer
        id_pedido:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        timestamp:
          type: string
    PedidoRequest:
      type: object
      properties:
        nome:
          type: string
        origem:
          type: string
        destino:
          type: string
        id_usuario:
          type: integer
        origemNome:
          type: string
        destinoNome:
          type: string
      required: [nome, origem, destino, id_usuario]
    PedidoUpdateRequest:
      type: object
      properties:
        nome:
          type: string
        origem:
          type: string
        destino:
          type: string
        situacao:
          type: integer
    Pedido:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        origem:
          type: string
        destino:
          type: string
        situacao:
          type: integer
        data_criacao:
          type: string
        id_usuario:
          type: integer
    RotaRequest:
      type: object
      properties:
        origem:
          type: string
        destino:
          type: string
      required: [origem, destino]
    RotaResponse:
      type: object
      properties:
        distancia:
          type: number
        tempo_estimado:
          type: integer
        coordenadas:
          type: array
          items:
            type: array
            items:
              type: number
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        senha:
          type: string
      required: [email, senha]
    RegisterRequest:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
        senha:
          type: string
        tipo:
          type: integer
      required: [nome, email, senha, tipo]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        usuario:
          type: object
    Usuario:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        tipo:
          type: integer
        data_criacao:
          type: string
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string